+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,25 @@
 # TBD
 
+# 0.20.0
+### Features
+* Added persistent volume claim creation to kubernetes-backed enclaves
+* Added `CreateEnclave` functionality to kubernetes backend
+* Added `ServiceAccounts`, `Roles`, `RoleBindings`, `ClusterRole`, and `ClusterRoleBindings` create, getByLabels and remove methods to `KubernetesManager`
+* Added `ForEngineNamespace`, `ForEngineServiceAccount`, `ForEngineClusterRole` and `ForEngineClusterRoleBindings` to  `KubernetesEngineObjectAttributesProvider`
+* Updated `KubernetesBackend.CreateEngine` added the kubernetes role based resources creation and namespace creation process
+* Fixed `KubernetesBackend.GetEngines`returning an empty list for filters with no IDs specified
+* Added a (currently unused) framework for collecting all Kubernetes resource that match a specific filter
+* Add `getEngineKubernetesResources` in preparation for refactoring the engine methods
+* Implement `KubernetesKurtosisBackend.DestroyEngines`
+
+### Changes
+* Updated `KubernetesManager.CreatePod` added `serviceAccount` argument to set the pod's service account
+* Switched all the engine methods to use a more Kubernetes-friendly way of getting & managing resources
+* Cleaned up the `KubernetesManager.CreateEngine` method
+
+### Breaking Changes
+* NewKurtosisKubernetesBackend now takes in extra arguments - `volumeStorageClassName` and `volumeSizePerEnclaveInGigabytes`
+
 # 0.19.0
 ### Breaking Changes
 * Removed `enclaveDataDirpathOnHostMachine` and `enclaveDataDirpathOnServiceContainer` from `KurtosisBackend.CreateUserService`